Add initial URL-matching logic for morph-url
See #8 for initial ideas around morph-url.

Lay down the URL-matching function for now. morphUrl is capable of
isolating URLs within a string and surrounding these URLs with anchor
tags.

Moving forward, we'll want to use the result of morphUrl to directly set
the HTML of an element from within a directive, to allow the anchor tags
to actually be rendered into the DOM.